pyproject.toml
[build-system]
requires = ["setuptools >= 61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cactus-test-definitions"
dynamic = ["version", "readme"]
authors = [{ name = "Mike Turner", email = "mike.turner@anu.edu.au" }]
maintainers = [{ name = "Mike Turner", email = "mike.turner@anu.edu.au" }]
description = "CSIP-AUS Client Test Harness Test Definitions"
keywords = ["CSIP-AUS", "client", "testing", "definitions"]
requires-python = ">=3.12"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development",
  "Typing :: Typed",
]
dependencies = [
  "pyyaml>=6.0.2,<7",
  "pyyaml-include>=2.2,<3",
  "dataclass-wizard==0.35.0,<1",
]

[project.urls]
Homepage = "https://github.com/bsgip/cactus-test-definitions"
Documentation = "https://github.com/bsgip/cactus-test-definitions/blob/main/README.md"
Repository = "https://github.com/bsgip/cactus-test-definitions.git"
Issues = "https://github.com/bsgip/cactus-test-definitions/issues"
Changelog = "https://github.com/bsgip/cactus-test-definitions/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
  "bandit",
  "black",
  "flake8",
  "isort",
  "mccabe",
  "mypy",
  "tox",
  "python-dotenv[cli]",
  "types-PyYAML",
]
test = ["pytest"]
docs = ["sphinx"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]


[tool.black]
line-length = 120

[tool.isort]
profile = "black"
src_paths = ["src", "tests"]

[tool.bandit]
exclude_dirs = ["tests"]

[tool.mypy]
exclude = ["dist", "build", "tests"]

[tool.setuptools.dynamic]
version = { attr = "cactus_test_definitions.__version__" }
readme = { file = ["README.md"], content-type = "text/markdown" }


[tool.setuptools.package-data]
"cactus_test_definitions" = ["py.typed"]
